fix(Interaction): ensure touch state is reset on disable
If an interactable object is touched and grabbed by one controller
but then another controller touches the object whilst it's grabbed
and then the object is disabled, the touch is not reset on the
original controller because it's not considered the touching obejct.

This fix ensures that the grabbing object and using object also
force the touch to be reset as well.